Move filename generation to own file
